requirements.txt
# =============================================================================
# 3579 Trading System - Core Dependencies
# =============================================================================
# Runtime dependencies for the trading system
# Install: pip install -r requirements.txt

# Core computation
numpy>=2.0.0

# Deep Learning (CPU version for CI, use torch+cu* for GPU)
torch>=2.0.0

# Data validation
pydantic>=2.0.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0.0

# HTTP client (for alerts/notifications)
requests>=2.28.0

# Output formatting
colorama>=0.4.0

# Build tools (for exe packaging)
pyinstaller>=6.0.0

# Tracing / Observability
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-instrumentation>=0.36b0
opentelemetry-instrumentation-fastapi>=0.36b0
opentelemetry-instrumentation-requests>=0.36b0
opentelemetry-exporter-otlp-proto-grpc>=1.19.0
opentelemetry-exporter-otlp-proto-http>=1.19.0
opentelemetry-exporter-console>=1.19.0
opentelemetry-instrumentation-azure-functions>=0.36b0
